Flyttet responsobjekter fra bidrag-grunnlag til egne filer. Fjernet overgangsstønad permanent.
